Add a QuantityUsed property to User.
In some licenses a user/host/display can use more than one license of a
feature. lmstat displays this with an addition to the user line of the
report. User now parses this addition and records the value in
QuantityUsed.

Change LmStatReportGenerator.LmStatTest to include examples of a user
that has more than one license checked out.

Update LicenseManager.Test to test the changes in Feature and User.